Update index.html
Changed title, changed how the matrix code behaves after clicking zero with awaiting period. Thought about adding a digit to be able to click before the wave settled but didn't seem to function first attempt. It keeps count of zeros and kind of splashes some of the numbers away from area of click.